Spatial consistency for surface movement  Particles are weighted by their residence time when measured on the surface.

+++ GULLI-Core/src/main/java/com/saemann/gulli/core/control/ShapeTools.java
@@ -373,7 +373,7 @@ public class ShapeTools {
                         //index out of bounds
                         continue;
                     }
-                    int[] particles = measurement.getParticlecount()[materialIndex];
+                    double[] particles = measurement.getParticlecount()[materialIndex];
                     int sum = 0;
                     for (int i = 0; i < particles.length; i++) {
                         sum += particles[i];
